+++ src/lib/output/plugins/index.ts
@@ -1,4 +1,3 @@
 export { AssetsPlugin } from "./AssetsPlugin";
 export { JavascriptIndexPlugin } from "./JavascriptIndexPlugin";
 export { MarkedPlugin } from "../themes/MarkedPlugin";
-export { LegendPlugin } from "./LegendPlugin";