+++ src/lib/output/themes/default/assets/typedoc/Theme.ts
@@ -1,13 +1,14 @@
+import { storage } from "./utils/storage";
+
 type ThemeChoice = "os" | "light" | "dark";
 
 export function initTheme(choices: HTMLOptionElement) {
-    const savedTheme =
-        (localStorage.getItem("tsd-theme") as ThemeChoice) || "os";
+    const savedTheme = (storage.getItem("tsd-theme") as ThemeChoice) || "os";
     choices.value = savedTheme;
     setTheme(savedTheme);
 
     choices.addEventListener("change", () => {
-        localStorage.setItem("tsd-theme", choices.value);
+        storage.setItem("tsd-theme", choices.value);
         setTheme(choices.value as ThemeChoice);
     });
 }